requirements.txt
babel==2.16.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
Flask==3.1.0
Flask-Login==0.6.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.1
numpy-financial==1.0.0
packaging==24.2
pillow==11.0.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
requests==2.32.3
setuptools==75.6.0
six==1.17.0
tkcalendar==1.6.1
tkinter-tooltip==3.1.2
ttkthemes==3.2.2
ttkwidgets==0.13.0
urllib3==2.3.0
Werkzeug==3.1.3
